requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
django-types==0.16.0
djangorestframework==3.14.0
djangorestframework-types==0.8.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
